services/jnsy-label/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
openpyxl==3.1.5
httpx>=0.24.0
